rename fieldParser.js to fieldsParser.js
