+++ commandParsers/fieldsParser.js
